requirements.txt
PyYAML==6.0.1
numpy==1.24.4
pillow==10.4.0
torch==2.3.1
matplotlib==3.7.5
torchvision==0.18.1
tqdm==4.66.5
fastapi==0.112.0
uvicorn==0.30.5
gunicorn